Clean up agent stream and OpenAI websocket plumbing
