+++ apps/os/src/domains/agents/durable-objects/agent-durable-object.ts
@@ -1,10 +1,5 @@
 import OpenAI from "openai";
-import type { ResponsesClientEvent } from "openai/resources/responses/responses";
-import { ResponsesWSBase } from "openai/resources/responses/ws-base";
-import { z } from "zod";
 import { DurableObject } from "cloudflare:workers";
-import type { ProcessorStreamApi } from "@iterate-com/shared/streams/stream-processors";
-import type { Event, EventInput, StreamCursor } from "@iterate-com/shared/streams/types";
 import { StreamPath } from "@iterate-com/shared/streams/types";
 import type { StreamEvent } from "@iterate-com/shared/streams/stream-event";
 import type { StreamRpc } from "~/domains/streams/engine/types.ts";
@@ -27,11 +22,8 @@ import {
   CloudflareAiProcessor,
   type CloudflareAiBinding,
 } from "~/domains/agents/stream-processors/cloudflare-ai/implementation.ts";
-import {
-  OpenAiWsProcessor,
-  type OpenAiResponsesWebSocket,
-  type OpenAiResponsesWebSocketStreamMessage,
-} from "~/domains/agents/stream-processors/openai-ws/implementation.ts";
+import { OpenAiWsProcessor } from "~/domains/agents/stream-processors/openai-ws/implementation.ts";
+import { createOpenAiResponsesWebSocketClient } from "~/domains/agents/stream-processors/openai-ws/cloudflare-responses-websocket.ts";
 import {
   getInitializedStreamStub,
   getStreamDurableObjectName,
@@ -44,7 +36,7 @@ import {
   type RepoInfo,
 } from "~/domains/repos/durable-objects/repo-durable-object.ts";
 import { getReposCapability } from "~/domains/repos/entrypoints/repo-capability.ts";
-import { resolveStreamPath } from "~/domains/streams/entrypoints/streams-backend.ts";
+import { getStreamsBackend } from "~/domains/streams/entrypoints/streams-backend.ts";
 import type { WorkspaceDurableObject } from "~/domains/workspaces/durable-objects/workspace-durable-object.ts";
 import { stripArtifactTokenQuery } from "~/domains/repos/artifact-token.ts";
 import {
@@ -87,23 +79,6 @@ export type CloneProjectRepoInput = {
   git: Awaited<ReturnType<WorkspaceDurableObject["cloudflareShellGit"]>>;
   repo: RepoInfo;
   workspace: DurableObjectStub<WorkspaceDurableObject>;
-};
-
-type AgentStreamApi = Omit<
-  ProcessorStreamApi<{
-    emits: readonly string[];
-    events: Record<string, unknown>;
-    processorDeps?: readonly unknown[];
-  }>,
-  "append" | "appendBatch" | "read"
-> & {
-  append(args: { event: EventInput; streamPath?: string }): Promise<Event>;
-  appendBatch(args: { events: EventInput[]; streamPath?: string }): Promise<Event[]>;
-  read(args?: {
-    streamPath?: string;
-    afterOffset?: StreamCursor;
-    beforeOffset?: StreamCursor;
-  }): Promise<Event[]>;
 };
 
 /** Bump when agentContextCapabilities changes — re-provides the agent's tools
@@ -713,10 +688,13 @@ export class AgentDurableObject extends DurableObject<AgentDurableObjectEnv> {
   }
 
   private streamsEntrypoint(streamPath: StreamPath) {
-    return agentStreamApiFromProject({
-      durableObjectNamespace: this.env.STREAM as unknown as StreamDurableObjectNamespace,
-      projectId: this.projectId(),
-      streamPath,
+    return getStreamsBackend({
+      exports: this.ctx.exports,
+      props: {
+        appendPolicy: { mode: "stream" },
+        projectId: this.projectId(),
+        streamPath: String(streamPath),
+      },
     });
   }
 }
@@ -734,76 +712,6 @@ export function readOpenAiApiKey(env: Record<string, unknown>) {
   } catch {
     return "";
   }
-}
-
-function agentStreamApiFromProject(args: {
-  durableObjectNamespace: StreamDurableObjectNamespace;
-  projectId: string;
-  streamPath: StreamPath;
-}): AgentStreamApi {
-  return {
-    async append(input) {
-      const stream = await getInitializedStreamStub({
-        durableObjectNamespace: args.durableObjectNamespace,
-        projectId: args.projectId,
-        path: resolveProcessorStreamPath({
-          basePath: args.streamPath,
-          pathInput: input.streamPath,
-        }),
-      });
-      return await stream.append(input.event);
-    },
-    async appendBatch(input) {
-      const stream = await getInitializedStreamStub({
-        durableObjectNamespace: args.durableObjectNamespace,
-        projectId: args.projectId,
-        path: resolveProcessorStreamPath({
-          basePath: args.streamPath,
-          pathInput: input.streamPath,
-        }),
-      });
-      return await stream.appendBatch(input.events);
-    },
-    async read(input = {}) {
-      const stream = await getInitializedStreamStub({
-        durableObjectNamespace: args.durableObjectNamespace,
-        projectId: args.projectId,
-        path: resolveProcessorStreamPath({
-          basePath: args.streamPath,
-          pathInput: input.streamPath,
-        }),
-      });
-      return await stream.history({
-        after: input.afterOffset,
-        before: input.beforeOffset ?? "end",
-      });
-    },
-    async *subscribe(input = {}) {
-      void input;
-      yield* [];
-      throw new Error("Agent processors receive live events through afterAppend RPC.");
-    },
-  };
-}
-
-function resolveProcessorStreamPath(input: { basePath: StreamPath; pathInput?: string }) {
-  if (input.pathInput == null) {
-    return input.basePath;
-  }
-
-  const trimmedPath = input.pathInput.trim();
-  if (!trimmedPath) {
-    throw new Error("Stream path is required.");
-  }
-
-  if (trimmedPath.startsWith("/")) {
-    return resolveStreamPath(trimmedPath);
-  }
-
-  const relativePath = trimmedPath.replace(/^\.\//, "").replace(/^\/+/, "");
-  return StreamPath.parse(
-    input.basePath === "/" ? `/${relativePath}` : `${input.basePath}/${relativePath}`,
-  );
 }
 
 /** AN AGENT IS A CONTEXT: its ref and address are projections of the
@@ -877,223 +785,4 @@ function parseChatToolMessage(value: unknown) {
 function isSlackAgentPath(agentPath: string) {
   const normalized = agentPath.toLowerCase();
   return normalized === "/agents/slack" || normalized.startsWith("/agents/slack/");
-}
-
-type CloudflareSocketEventName = "open" | "message" | "close" | "error" | string;
-type JsonValue = z.infer<ReturnType<typeof z.json>>;
-
-export function createOpenAiResponsesWebSocketClient(client: OpenAI): OpenAiResponsesWebSocket {
-  const sdkWebSocket = new CloudflareResponsesWebSocket(client);
-
-  return {
-    get url() {
-      return sdkWebSocket.url;
-    },
-    get socket() {
-      return sdkWebSocket.socket;
-    },
-    send(event) {
-      sdkWebSocket.send(event as unknown as ResponsesClientEvent);
-    },
-    stream() {
-      return streamOpenAiResponsesWebSocket(sdkWebSocket);
-    },
-    close(props) {
-      sdkWebSocket.close(props);
-    },
-  };
-}
-
-async function* streamOpenAiResponsesWebSocket(
-  sdkWebSocket: CloudflareResponsesWebSocket,
-): AsyncIterableIterator<OpenAiResponsesWebSocketStreamMessage> {
-  for await (const event of sdkWebSocket.stream()) {
-    switch (event.type) {
-      case "connecting":
-      case "open":
-      case "closing":
-      case "reconnected":
-        yield { type: event.type };
-        break;
-      case "close":
-        yield { type: "close", code: event.code, reason: event.reason };
-        break;
-      case "reconnecting":
-        yield { type: "reconnecting", reconnect: toJsonValue(event.reconnect) };
-        break;
-      case "message":
-        yield { type: "message", message: toJsonValue(event.message) };
-        break;
-      case "raw":
-        yield { type: "raw", data: event.data };
-        break;
-      case "error":
-        yield { type: "error", error: event.error };
-        break;
-      default:
-        event satisfies never;
-    }
-  }
-}
-
-function toJsonValue(value: unknown): JsonValue {
-  return z.json().parse(value);
-}
-
-class CloudflareResponsesWebSocket extends ResponsesWSBase<CloudflareFetchWebSocket> {
-  constructor(client: OpenAI) {
-    super(client, { reconnect: null });
-    this._connectInitial();
-  }
-
-  protected _createSocket(url: URL, authHeaders: Record<string, string>): CloudflareFetchWebSocket {
-    return new CloudflareFetchWebSocket(url, {
-      ...authHeaders,
-      "OpenAI-Beta": "responses_websockets=2026-02-06",
-    });
-  }
-}
-
-class CloudflareFetchWebSocket {
-  #listeners = new Map<CloudflareSocketEventName, Set<unknown>>();
-  #onceListeners = new Map<CloudflareSocketEventName, Map<unknown, unknown>>();
-  #readyState = 0;
-  #socket: WebSocket | undefined;
-
-  constructor(
-    private readonly url: URL,
-    private readonly authHeaders: Record<string, string>,
-  ) {
-    void this.#connect();
-  }
-
-  get readyState(): number {
-    return this.#socket?.readyState ?? this.#readyState;
-  }
-
-  send(data: string | ArrayBufferLike | ArrayBufferView): void {
-    if (this.#socket == null) throw new Error("OpenAI WebSocket is not open.");
-    this.#socket.send(data);
-  }
-
-  close(code?: number, reason?: string): void {
-    this.#readyState = 2;
-    this.#socket?.close(code, reason);
-  }
-
-  on(event: "open", listener: () => void): void;
-  on(
-    event: "message",
-    listener: (data: string | ArrayBuffer | ArrayBufferView, isBinary: boolean) => void,
-  ): void;
-  on(event: "close", listener: (code: number, reason: string) => void): void;
-  on(event: "error", listener: (error: Error) => void): void;
-  on(event: CloudflareSocketEventName, listener: (...args: never[]) => void): void;
-  on(event: CloudflareSocketEventName, listener: unknown): void {
-    this.#listenersFor(event).add(listener);
-  }
-
-  off(event: "open", listener: () => void): void;
-  off(
-    event: "message",
-    listener: (data: string | ArrayBuffer | ArrayBufferView, isBinary: boolean) => void,
-  ): void;
-  off(event: "close", listener: (code: number, reason: string) => void): void;
-  off(event: "error", listener: (error: Error) => void): void;
-  off(event: CloudflareSocketEventName, listener: (...args: never[]) => void): void;
-  off(event: CloudflareSocketEventName, listener: unknown): void {
-    this.#removeListener(event, listener);
-  }
-
-  once(event: "open", listener: () => void): void;
-  once(
-    event: "message",
-    listener: (data: string | ArrayBuffer | ArrayBufferView, isBinary: boolean) => void,
-  ): void;
-  once(event: "close", listener: (code: number, reason: string) => void): void;
-  once(event: "error", listener: (error: Error) => void): void;
-  once(event: CloudflareSocketEventName, listener: (...args: never[]) => void): void;
-  once(event: CloudflareSocketEventName, listener: unknown): void {
-    const onceListener = (...args: never[]) => {
-      this.#removeListener(event, listener);
-      (listener as (...args: never[]) => void)(...args);
-    };
-    this.#onceListenersFor(event).set(listener, onceListener);
-    this.on(event, onceListener);
-  }
-
-  get socket(): { readonly readyState: number } {
-    return { readyState: this.readyState };
-  }
-
-  async #connect() {
-    try {
-      const response = (await fetch(this.url.toString().replace("wss://", "https://"), {
-        headers: {
-          ...this.authHeaders,
-          Upgrade: "websocket",
-        },
-      })) as Response & { webSocket?: WebSocket | null };
-
-      if (response.webSocket == null) {
-        throw new Error(`OpenAI WebSocket upgrade failed with status ${response.status}.`);
-      }
-
-      this.#socket = response.webSocket;
-      this.#socket.accept();
-      this.#bindSocket(this.#socket);
-      this.#readyState = this.#socket.readyState;
-      this.#emit("open");
-    } catch (error) {
-      this.#readyState = 3;
-      this.#emit("error", error instanceof Error ? error : new Error(String(error)));
-      this.#emit("close", 1006, "OpenAI WebSocket upgrade failed.");
-    }
-  }
-
-  #bindSocket(socket: WebSocket) {
-    socket.addEventListener("message", (event) => {
-      this.#emit("message", event.data, event.data instanceof ArrayBuffer);
-    });
-    socket.addEventListener("close", (event) => {
-      this.#readyState = 3;
-      this.#emit("close", event.code, event.reason);
-    });
-    socket.addEventListener("error", () => {
-      this.#emit("error", new Error("OpenAI WebSocket errored."));
-    });
-  }
-
-  #listenersFor(event: CloudflareSocketEventName): Set<unknown> {
-    const existing = this.#listeners.get(event);
-    if (existing != null) return existing;
-
-    const listeners = new Set<unknown>();
-    this.#listeners.set(event, listeners);
-    return listeners;
-  }
-
-  #onceListenersFor(event: CloudflareSocketEventName): Map<unknown, unknown> {
-    const existing = this.#onceListeners.get(event);
-    if (existing != null) return existing;
-
-    const listeners = new Map<unknown, unknown>();
-    this.#onceListeners.set(event, listeners);
-    return listeners;
-  }
-
-  #removeListener(event: CloudflareSocketEventName, listener: unknown) {
-    const listeners = this.#listeners.get(event);
-    listeners?.delete(listener);
-    const onceListener = this.#onceListeners.get(event)?.get(listener);
-    if (onceListener == null) return;
-    listeners?.delete(onceListener);
-    this.#onceListeners.get(event)?.delete(listener);
-  }
-
-  #emit(event: CloudflareSocketEventName, ...args: unknown[]) {
-    for (const listener of this.#listeners.get(event) ?? []) {
-      (listener as (...args: unknown[]) => void)(...args);
-    }
-  }
 }
