+++ apps/os/src/domains/agents/stream-processors/openai-ws/cloudflare-responses-websocket.ts
@@ -1,0 +1,227 @@
+import OpenAI from "openai";
+import type { ResponsesClientEvent } from "openai/resources/responses/responses";
+import { ResponsesWSBase } from "openai/resources/responses/ws-base";
+import { z } from "zod";
+import type {
+  OpenAiResponsesWebSocket,
+  OpenAiResponsesWebSocketStreamMessage,
+} from "./implementation.ts";
+
+type CloudflareSocketEventName = "open" | "message" | "close" | "error" | string;
+type JsonValue = z.infer<ReturnType<typeof z.json>>;
+
+export function createOpenAiResponsesWebSocketClient(client: OpenAI): OpenAiResponsesWebSocket {
+  const sdkWebSocket = new CloudflareResponsesWebSocket(client);
+
+  return {
+    get url() {
+      return sdkWebSocket.url;
+    },
+    get socket() {
+      return sdkWebSocket.socket;
+    },
+    send(event) {
+      sdkWebSocket.send(event as unknown as ResponsesClientEvent);
+    },
+    stream() {
+      return streamOpenAiResponsesWebSocket(sdkWebSocket);
+    },
+    close(props) {
+      sdkWebSocket.close(props);
+    },
+  };
+}
+
+async function* streamOpenAiResponsesWebSocket(
+  sdkWebSocket: CloudflareResponsesWebSocket,
+): AsyncIterableIterator<OpenAiResponsesWebSocketStreamMessage> {
+  for await (const event of sdkWebSocket.stream()) {
+    switch (event.type) {
+      case "connecting":
+      case "open":
+      case "closing":
+      case "reconnected":
+        yield { type: event.type };
+        break;
+      case "close":
+        yield { type: "close", code: event.code, reason: event.reason };
+        break;
+      case "reconnecting":
+        yield { type: "reconnecting", reconnect: toJsonValue(event.reconnect) };
+        break;
+      case "message":
+        yield { type: "message", message: toJsonValue(event.message) };
+        break;
+      case "raw":
+        yield { type: "raw", data: event.data };
+        break;
+      case "error":
+        yield { type: "error", error: event.error };
+        break;
+      default:
+        event satisfies never;
+    }
+  }
+}
+
+function toJsonValue(value: unknown): JsonValue {
+  return z.json().parse(value);
+}
+
+class CloudflareResponsesWebSocket extends ResponsesWSBase<CloudflareFetchWebSocket> {
+  constructor(client: OpenAI) {
+    super(client, { reconnect: null });
+    this._connectInitial();
+  }
+
+  protected _createSocket(url: URL, authHeaders: Record<string, string>): CloudflareFetchWebSocket {
+    return new CloudflareFetchWebSocket(url, {
+      ...authHeaders,
+      "OpenAI-Beta": "responses_websockets=2026-02-06",
+    });
+  }
+}
+
+class CloudflareFetchWebSocket {
+  #listeners = new Map<CloudflareSocketEventName, Set<unknown>>();
+  #onceListeners = new Map<CloudflareSocketEventName, Map<unknown, unknown>>();
+  #readyState = 0;
+  #socket: WebSocket | undefined;
+
+  constructor(
+    private readonly url: URL,
+    private readonly authHeaders: Record<string, string>,
+  ) {
+    void this.#connect();
+  }
+
+  get readyState(): number {
+    return this.#socket?.readyState ?? this.#readyState;
+  }
+
+  send(data: string | ArrayBufferLike | ArrayBufferView): void {
+    if (this.#socket == null) throw new Error("OpenAI WebSocket is not open.");
+    this.#socket.send(data);
+  }
+
+  close(code?: number, reason?: string): void {
+    this.#readyState = 2;
+    this.#socket?.close(code, reason);
+  }
+
+  on(event: "open", listener: () => void): void;
+  on(
+    event: "message",
+    listener: (data: string | ArrayBuffer | ArrayBufferView, isBinary: boolean) => void,
+  ): void;
+  on(event: "close", listener: (code: number, reason: string) => void): void;
+  on(event: "error", listener: (error: Error) => void): void;
+  on(event: CloudflareSocketEventName, listener: (...args: never[]) => void): void;
+  on(event: CloudflareSocketEventName, listener: unknown): void {
+    this.#listenersFor(event).add(listener);
+  }
+
+  off(event: "open", listener: () => void): void;
+  off(
+    event: "message",
+    listener: (data: string | ArrayBuffer | ArrayBufferView, isBinary: boolean) => void,
+  ): void;
+  off(event: "close", listener: (code: number, reason: string) => void): void;
+  off(event: "error", listener: (error: Error) => void): void;
+  off(event: CloudflareSocketEventName, listener: (...args: never[]) => void): void;
+  off(event: CloudflareSocketEventName, listener: unknown): void {
+    this.#removeListener(event, listener);
+  }
+
+  once(event: "open", listener: () => void): void;
+  once(
+    event: "message",
+    listener: (data: string | ArrayBuffer | ArrayBufferView, isBinary: boolean) => void,
+  ): void;
+  once(event: "close", listener: (code: number, reason: string) => void): void;
+  once(event: "error", listener: (error: Error) => void): void;
+  once(event: CloudflareSocketEventName, listener: (...args: never[]) => void): void;
+  once(event: CloudflareSocketEventName, listener: unknown): void {
+    const onceListener = (...args: never[]) => {
+      this.#removeListener(event, listener);
+      (listener as (...args: never[]) => void)(...args);
+    };
+    this.#onceListenersFor(event).set(listener, onceListener);
+    this.on(event, onceListener);
+  }
+
+  get socket(): { readonly readyState: number } {
+    return { readyState: this.readyState };
+  }
+
+  async #connect() {
+    try {
+      const response = (await fetch(this.url.toString().replace("wss://", "https://"), {
+        headers: {
+          ...this.authHeaders,
+          Upgrade: "websocket",
+        },
+      })) as Response & { webSocket?: WebSocket | null };
+
+      if (response.webSocket == null) {
+        throw new Error(`OpenAI WebSocket upgrade failed with status ${response.status}.`);
+      }
+
+      this.#socket = response.webSocket;
+      this.#socket.accept();
+      this.#bindSocket(this.#socket);
+      this.#readyState = this.#socket.readyState;
+      this.#emit("open");
+    } catch (error) {
+      this.#readyState = 3;
+      this.#emit("error", error instanceof Error ? error : new Error(String(error)));
+      this.#emit("close", 1006, "OpenAI WebSocket upgrade failed.");
+    }
+  }
+
+  #bindSocket(socket: WebSocket) {
+    socket.addEventListener("message", (event) => {
+      this.#emit("message", event.data, event.data instanceof ArrayBuffer);
+    });
+    socket.addEventListener("close", (event) => {
+      this.#readyState = 3;
+      this.#emit("close", event.code, event.reason);
+    });
+    socket.addEventListener("error", () => {
+      this.#emit("error", new Error("OpenAI WebSocket errored."));
+    });
+  }
+
+  #listenersFor(event: CloudflareSocketEventName): Set<unknown> {
+    const existing = this.#listeners.get(event);
+    if (existing != null) return existing;
+
+    const listeners = new Set<unknown>();
+    this.#listeners.set(event, listeners);
+    return listeners;
+  }
+
+  #onceListenersFor(event: CloudflareSocketEventName): Map<unknown, unknown> {
+    const existing = this.#onceListeners.get(event);
+    if (existing != null) return existing;
+
+    const listeners = new Map<unknown, unknown>();
+    this.#onceListeners.set(event, listeners);
+    return listeners;
+  }
+
+  #removeListener(event: CloudflareSocketEventName, listener: unknown) {
+    const listeners = this.#listeners.get(event);
+    listeners?.delete(listener);
+    const onceListener = this.#onceListeners.get(event)?.get(listener);
+    if (onceListener == null) return;
+    listeners?.delete(onceListener);
+    this.#onceListeners.get(event)?.delete(listener);
+  }
+
+  #emit(event: CloudflareSocketEventName, ...args: unknown[]) {
+    for (const listener of this.#listeners.get(event) ?? []) {
+      (listener as (...args: unknown[]) => void)(...args);
+    }
+  }
+}
